Remove SpanExporter from Jaeger
renaming files and updating references

merging JaegerExporter and JaegerTraceExporter

merging tests

update
