+++ src/OpenTelemetry.Exporter.Jaeger/JaegerExporter.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="JaegerActivityExporter.cs" company="OpenTelemetry Authors">
+﻿// <copyright file="JaegerExporter.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,19 +16,23 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using OpenTelemetry.Exporter.Jaeger.Implementation;
+using OpenTelemetry.Resources;
+using OpenTelemetry.Trace;
 using OpenTelemetry.Trace.Export;
 
 namespace OpenTelemetry.Exporter.Jaeger
 {
-    public class JaegerActivityExporter : ActivityExporter, IDisposable
+    public class JaegerExporter : ActivityExporter, IDisposable
     {
         private readonly SemaphoreSlim exportLock = new SemaphoreSlim(1);
+        private bool libraryResourceApplied = false;
         private bool disposedValue = false; // To detect redundant dispose calls
 
-        public JaegerActivityExporter(JaegerExporterOptions options)
+        public JaegerExporter(JaegerExporterOptions options)
         {
             this.JaegerAgentUdpBatcher = new JaegerUdpBatcher(options);
         }
@@ -40,6 +44,15 @@ namespace OpenTelemetry.Exporter.Jaeger
             await this.exportLock.WaitAsync().ConfigureAwait(false);
             try
             {
+                if (!this.libraryResourceApplied && activityBatch.Count() > 0)
+                {
+                    var libraryResource = activityBatch.First().GetResource();
+
+                    this.ApplyLibraryResource(libraryResource ?? Resource.Empty);
+
+                    this.libraryResourceApplied = true;
+                }
+
                 foreach (var activity in activityBatch)
                 {
                     // TODO: group by activity source
@@ -67,6 +80,58 @@ namespace OpenTelemetry.Exporter.Jaeger
         {
             // Do not change this code. Put cleanup code in Dispose(bool disposing).
             this.Dispose(true);
+        }
+
+        internal void ApplyLibraryResource(Resource libraryResource)
+        {
+            if (libraryResource is null)
+            {
+                throw new ArgumentNullException(nameof(libraryResource));
+            }
+
+            var process = this.JaegerAgentUdpBatcher.Process;
+
+            string serviceName = null;
+            string serviceNamespace = null;
+            foreach (var label in libraryResource.Attributes)
+            {
+                string key = label.Key;
+
+                if (label.Value is string strVal)
+                {
+                    switch (key)
+                    {
+                        case Resource.ServiceNameKey:
+                            serviceName = strVal;
+                            continue;
+                        case Resource.ServiceNamespaceKey:
+                            serviceNamespace = strVal;
+                            continue;
+                        case Resource.LibraryNameKey:
+                        case Resource.LibraryVersionKey:
+                            continue;
+                    }
+                }
+
+                if (process.Tags == null)
+                {
+                    process.Tags = new Dictionary<string, JaegerTag>();
+                }
+
+                process.Tags[label.Key] = label.ToJaegerTag();
+            }
+
+            if (serviceName != null)
+            {
+                process.ServiceName = serviceNamespace != null
+                    ? serviceNamespace + "." + serviceName
+                    : serviceName;
+            }
+
+            if (string.IsNullOrEmpty(process.ServiceName))
+            {
+                process.ServiceName = JaegerExporterOptions.DefaultServiceName;
+            }
         }
 
         protected virtual void Dispose(bool disposing)
